Add project files.

+++ Practicum.API/Controllers/ProductsController.cs
@@ -23,21 +23,23 @@ namespace Practicum.API.Controllers
         {
             var products = await _service.GetAllAsync();
             var productsDtos = _mapper.Map<List<ProductDto>>(products.ToList());
-            return CreateActionResult(CustomResponseDto<List<ProductDto>>.Success(200, productsDtos));
+            return Ok(CustomResponseDto<List<ProductDto>>.Success(200, productsDtos));
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
             var product = await _service.GetByIdAsync(id);
+            if (product == null)
+                return NotFound();
             var productsDtos = _mapper.Map<ProductDto>(product);
-            return CreateActionResult(CustomResponseDto<ProductDto>.Success(200, productsDtos));
+            return Ok(CustomResponseDto<ProductDto>.Success(200, productsDtos));
         }
 
         [HttpGet("[action]/{name}")]
         public async Task<IActionResult> GetByNameAsync(string name)
         {
-            return CreateActionResult(await _service.GetByNameAsync(name));
+            return Ok(await _service.GetByNameAsync(name));
         }
 
         [HttpPost]
@@ -45,22 +47,26 @@ namespace Practicum.API.Controllers
         {
             var product = await _service.AddAsync(_mapper.Map<Product>(productDto));
             var productsDtos = _mapper.Map<ProductDto>(product);
-            return CreateActionResult(CustomResponseDto<ProductDto>.Success(201, productsDtos));
+            return Ok(CustomResponseDto<ProductDto>.Success(201, productsDtos));
         }
 
         [HttpPut]
         public async Task<IActionResult> Update(ProductDto productDto)
         {
             await _service.UpdateAsync(_mapper.Map<Product>(productDto));
-            return CreateActionResult(CustomResponseDto<NoContentDto>.Success(204));
+            return Ok(CustomResponseDto<NoContentDto>.Success(204));
         }
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Remove(int id)
         {
             var product = await _service.GetByIdAsync(id);
+            if (product == null)
+            {
+                return NotFound();
+            }
             await _service.RemoveAsync(product);
-            return CreateActionResult(CustomResponseDto<NoContentDto>.Success(204));
+            return Ok(CustomResponseDto<NoContentDto>.Success(204));
         }
     }
 }
